Moved classes our of utils
