+++ src/main/java/io/strimzi/kafka/bridge/mqtt/kafka/KafkaProducerAckLevel.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.kafka.bridge.mqtt.utils;
+package io.strimzi.kafka.bridge.mqtt.kafka;
 
 /**
  * Represents the Kafka producer ack level
