+++ src/main/java/io/strimzi/kafka/bridge/mqtt/mapper/MappingRulesLoader.java
@@ -2,10 +2,9 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.kafka.bridge.mqtt.utils;
+package io.strimzi.kafka.bridge.mqtt.mapper;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.strimzi.kafka.bridge.mqtt.mapper.MappingRule;
 
 import java.io.IOException;
 import java.nio.file.Path;
